Fix build. NGT-2091

+++ src/core/lib/core_data_model/reflection/reflected_object_item_new.hpp
@@ -37,7 +37,7 @@ public:
 
 	virtual ReflectedTreeItemNew * getChild( size_t index ) const override;
 
-	virtual size_t size() const override;
+	virtual int size() const override;
 
 	virtual bool preSetValue( const PropertyAccessor & accessor,
 		const Variant & value ) override;
